1.0增加前端页面参数字符个数限制，如报名地点不能超过100个字符串

+++ gzxant-web/src/main/java/com/gzxant/controller/enroll/personnel/EnrollPersonnelController.java
@@ -162,7 +162,9 @@ public class EnrollPersonnelController extends BaseController {
         for(EnrollEnter list:listEnter){
             enterIds.add(list.getId());
         }
-        enrollEnterService.deleteBatchIds(enterIds);
+        if(enterIds.size()>0){
+			enrollEnterService.deleteBatchIds(enterIds);
+		}
 
 		boolean success = enrollPersonnelService.deleteBatchIds(ids);
 		if (success) {
